Text corrections
- About Norwegian Wood (This Bird Has Flown)
Assuming that `It's time for bed` is the only word she clearly said, the rest of the words are all "my" words.
I feel that translating `wood` as `家具` is a bit intrusive, so I'll just use the word `wood`.

Hmmm. I wonder😣

(I thought these comments might be interesting once in a while.)