Nerd Fonts v3.0.0 Update!!
This is a major release and they have made major changes.

refs: https://github.com/ryanoasis/nerd-fonts/releases/tag/v3.0.0

The icons used in the articles on the site have changed, but that's because we couldn't find the same font we were using before!
Maybe if I look for it properly, I can find it. 😅

But it's a good opportunity and I'll enjoy the change! 😆