fix: Zooming and scaling must not be disabled
Remove the `user-scalable="no"` parameter from the `content` attribute of the
`<meta name="viewport">` element in order to allow zooming and ensure the `maximum-scale` parameter is not less than 2.

A meta viewport element tells the browser how to control the page's dimensions and to scale,
but the ability to zoom can be disabled as a result of its use.

refs: https://dequeuniversity.com/rules/axe/4.6/meta-viewport

I was setting this because I have a problem with automatic zooming when entering form on the iPhone,
but it does not zoom back in after the user finishes typing.

This setting does not disallow manual pinch zooming by the user.

So I am frankly not convinced... 😟