Setting of external links is handled by `js` 📮
For a long time, I had been using wasm to handle the addition of external link settings,
but this was done in the first place for the purpose of learning and checking how wasm worked.

We have not measured the processing speed, but this is more advantageous in terms of file size.