Fixing the gears that weren't quite meshing Part2 ⚙️⚙️
This commit fixes an issue where, if a `video` exists within a `slider`, the sizes are not correctly reflected (until the video is actually played) if a `poster` is missing.

(In reality, simply using `preload=“metadata”` resolved the issue, but in Safari, video playback remains unstable unless set to “none”...)

It's a difficult decision, but this site will adhere to the `Plyr` standard and uniformly use only the `data-poster` attribute of the `video` element.